requirements.txt
langchain
langchain_community
langchain_groq
fastapi
streamlit
python-dotenv